initial release of objectSpawnSystem
This script allows you to set cells to persist through resets as when you saved them.
Example, lets say you decorate a cell in balmora in game, and always want the cell to look like that, even after the cell file is deleted. Using this feature to save the cell will make that cell always look like when you saved it whenever the cell is missing the saved objects on a reset.